Add MATLAB  gantry control App
Developer notes

Work on the MATLAB control program began around the same time as the
ClearCore code when I took over this project on Aug 11. It's the 2nd
part of the one whole thing. This commit includes work over two months.
Initially it was quick getting to control the carriage interactively
using a game controller. But to be able to do pathtracking and run two
carriages in coordination took much more time. A lot of the GUI code
was generated by AI agents.

The app was developed was developed bottom up. Each component can be
run individually. The main App, Jakiro, contains two instances of the
CarriageApp, one for each carriage, and one instance of the wavi class
for data aquisition. The CarriageApp has three main components: a serial
port manager (SerialPortApp), a mode/pathdata selector (ControlModePanel),
and a carriage controller (CarriageControl) for reading status and
sending commands to the ClearCore microcontroller.

Current status

Run the Jakiro class to start the control app.
The two carriages can do coordinated pathtracking for experiment runs,
with prescribed speed.

Todo

Integrate data collection in the experiment run.