AOA control
Previously AOA oscillates during path tracking. It was caused by oscillation of the target position on the path.

This is avoided by limiting target position from moving back.
For path tracking, using tangent angle of the path instead of instantaneous velocity. This is more stable but less truthful. But because of stuttering in the gantry system, velocity direction could have large swings.